Fix the reload timers not working correctly on magreload and main reload.
The mag reload now will retain its timer no matter the switching of the weapon or exiting of the vehicle, removing the quick swap from and back to weapon to reload instantly bug.

The main reload now will not allow you to skip the reload on all vehicles meaning that you now need to wait the correct time on a weapon after loading it before you can shoot again. Patching the glitch with the HIMARS allowing you to shoot continuously as long as you had the resources. Also fixes the jet spam when on ground in creative etc.