Update to stop the global client sound of Lockons
Updated to block the global lock-on sound, now players a different lock-on sound but it no longer is global? we will see if it works.

Updates to the way we call the sound now purely on peoples clients.

Might still need work later down the line + changes to how loud the sound plays maybe?